+++ packages/flutter_adaptive_cards/lib/src/inputs/text.dart
@@ -66,6 +66,9 @@ class AdaptiveTextInputState extends State<AdaptiveTextInput>
             SizedBox(
               height: 40,
               child: TextFormField(
+                key: ValueKey(
+                  '${(widget.key! as ValueKey<String>).value}_input',
+                ),
                 style: const TextStyle(),
                 controller: controller,
                 // maxLength: maxLength,