+++ packages/flutter_adaptive_cards/CHANGELOG.md
@@ -23,7 +23,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Addeed SVG support for images - but not background images
 - Added Inline in-memory image support - png
 - All elements have repeatable id values if not specified in the json (hashcode(adaptiveMap))
--
+- Changed action path again (simplified)
+- Put provider back in for raw and adaptive card element
+- Added keys for the Adaptive cards
+- Added calcuated and specified ids
+- Added keys to the input fields in the input adaptive cards and atual input widgets for testing
 
 ## [0.2.0](https://github.com/freemansoft/Flutter-AdaptiveCards/compare/0.1.2...0.2.0) - 2023-??-??